Re-organize existing tests to their own SettingsSearchTest class
