+++ app/src/androidTest/java/org/mozilla/fenix/ui/SmokeTest.kt
@@ -484,86 +484,6 @@ class SmokeTest {
     }
 
     @Test
-    // Ads a new search engine from the list of custom engines
-    fun addPredefinedSearchEngineTest() {
-        val searchEngine = "Reddit"
-
-        homeScreen {
-        }.openThreeDotMenu {
-        }.openSettings {
-        }.openSearchSubMenu {
-            openAddSearchEngineMenu()
-            verifyAddSearchEngineList()
-            addNewSearchEngine(searchEngine)
-            verifyEngineListContains(searchEngine)
-        }.goBack {
-        }.goBack {
-        }.openSearch {
-            verifyKeyboardVisibility()
-            clickSearchEngineShortcutButton()
-            verifyEnginesListShortcutContains(activityTestRule, searchEngine)
-            changeDefaultSearchEngine(activityTestRule, searchEngine)
-        }.submitQuery("mozilla ") {
-            verifyUrl(searchEngine)
-        }
-    }
-
-    @Test
-    // Verifies setting as default a customized search engine name and URL
-    @Ignore("Failing intermittently https://github.com/mozilla-mobile/fenix/issues/22256")
-    fun editCustomSearchEngineTest() {
-        val searchEngine = object {
-            var title = "Elefant"
-            var url = "https://www.elefant.ro/search?SearchTerm=%s"
-            var newTitle = "Test"
-        }
-
-        homeScreen {
-        }.openThreeDotMenu {
-        }.openSettings {
-        }.openSearchSubMenu {
-            openAddSearchEngineMenu()
-            selectAddCustomSearchEngine()
-            typeCustomEngineDetails(searchEngine.title, searchEngine.url)
-            saveNewSearchEngine()
-            openEngineOverflowMenu(searchEngine.title)
-            clickEdit()
-            typeCustomEngineDetails(searchEngine.newTitle, searchEngine.url)
-            saveEditSearchEngine()
-            changeDefaultSearchEngine(searchEngine.newTitle)
-        }.goBack {
-        }.goBack {
-        }.openSearch {
-            verifyDefaultSearchEngine(searchEngine.newTitle)
-            clickSearchEngineShortcutButton()
-            verifyEnginesListShortcutContains(activityTestRule, searchEngine.newTitle)
-        }
-    }
-
-    @Test
-    // Test running on beta/release builds in CI:
-    // caution when making changes to it, so they don't block the builds
-    // Goes through the settings and changes the search suggestion toggle, then verifies it changes.
-    fun toggleSearchSuggestions() {
-
-        homeScreen {
-        }.openSearch {
-            typeSearch("mozilla")
-            verifySearchEngineSuggestionResults(activityTestRule, "mozilla firefox")
-        }.dismissSearchBar {
-        }.openThreeDotMenu {
-        }.openSettings {
-        }.openSearchSubMenu {
-            disableShowSearchSuggestions()
-        }.goBack {
-        }.goBack {
-        }.openSearch {
-            typeSearch("mozilla")
-            verifyNoSuggestionsAreDisplayed(activityTestRule, "mozilla firefox")
-        }
-    }
-
-    @Test
     // Swipes the nav bar left/right to switch between tabs
     fun swipeToSwitchTabTest() {
         val firstWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
