+++ automation/taskcluster/androidTest/flank-x86-start-test.yml
@@ -37,7 +37,7 @@ gcloud:
    - class org.mozilla.fenix.ui.NavigationToolbarTest#visitURLTest
    - class org.mozilla.fenix.ui.HistoryTest#visitedUrlHistoryTest
    - class org.mozilla.fenix.ui.SmokeTest#openMainMenuSettingsItemTest
-   - class org.mozilla.fenix.ui.SmokeTest#toggleSearchSuggestions
+   - class org.mozilla.fenix.ui.SettingsSearchTest#toggleSearchSuggestionsTest
    - class org.mozilla.fenix.ui.SmokeTest#deleteCollectionTest
    - class org.mozilla.fenix.ui.SmokeTest#noHistoryInPrivateBrowsingTest
    - class org.mozilla.fenix.ui.NoNetworkAccessStartupTests#noNetworkConnectionStartupTest
